+++ common/src/main/java/foundry/veil/api/client/render/rendertype/VeilRenderType.java
@@ -72,37 +72,37 @@ public final class VeilRenderType extends RenderType {
                 .createCompositeState(false);
         return create(Veil.MODID + ":quasar_particle", VeilVertexFormat.QUASAR_PARTICLE, VertexFormat.Mode.QUADS, SMALL_BUFFER_SIZE, false, false, state);
     });
-    private static final Function<ResourceLocation, RenderType> QUASAR_TRAIL = Util.memoize((texture) -> {
+    private static final BiFunction<ResourceLocation, Boolean, RenderType> QUASAR_TRAIL = Util.memoize((texture, additive) -> {
         CompositeState state = CompositeState.builder()
                 .setShaderState(RENDERTYPE_ENTITY_TRANSLUCENT_EMISSIVE_SHADER)
                 .setTextureState(new TextureStateShard(texture, false, false))
-                .setTransparencyState(ADDITIVE_TRANSPARENCY)
+                .setTransparencyState(additive ? ADDITIVE_TRANSPARENCY : TRANSLUCENT_TRANSPARENCY)
                 .setWriteMaskState(COLOR_DEPTH_WRITE)
                 .setCullState(NO_CULL)
                 .createCompositeState(false);
         return RenderType.create(Veil.MODID + ":quasar_trail", DefaultVertexFormat.NEW_ENTITY, VertexFormat.Mode.TRIANGLE_STRIP, TRANSIENT_BUFFER_SIZE, false, false, state);
     });
-    private static final RenderType NO_TEXTURE_QUASAR_TRAIL = RenderType.create(
-            Veil.MODID + ":quasar_trail",
-            DefaultVertexFormat.NEW_ENTITY,
-            VertexFormat.Mode.TRIANGLE_STRIP,
-            TRANSIENT_BUFFER_SIZE,
-            false,
-            false,
-            CompositeState.builder()
-                    .setShaderState(RENDERTYPE_ENTITY_TRANSLUCENT_EMISSIVE_SHADER)
-                    .setTextureState(WHITE_TEXTURE)
-                    .setTransparencyState(ADDITIVE_TRANSPARENCY)
-                    .setWriteMaskState(COLOR_DEPTH_WRITE)
-                    .setCullState(NO_CULL)
-                    .createCompositeState(false));
+    private static final Function<Boolean, RenderType> NO_TEXTURE_QUASAR_TRAIL = Util.memoize((additive) -> {
+        CompositeState state = CompositeState.builder()
+                .setShaderState(RENDERTYPE_ENTITY_TRANSLUCENT_EMISSIVE_SHADER)
+                .setTextureState(WHITE_TEXTURE)
+                .setTransparencyState(additive ? ADDITIVE_TRANSPARENCY : TRANSLUCENT_TRANSPARENCY)
+                .setWriteMaskState(COLOR_DEPTH_WRITE)
+                .setCullState(NO_CULL)
+                .createCompositeState(false);
+        return RenderType.create(Veil.MODID + ":quasar_trail", DefaultVertexFormat.NEW_ENTITY, VertexFormat.Mode.TRIANGLE_STRIP, TRANSIENT_BUFFER_SIZE, false, false, state);
+    }) ;
 
     public static RenderType quasarParticle(ResourceLocation texture, boolean additive) {
         return QUASAR_PARTICLE.apply(texture, additive);
     }
 
     public static RenderType quasarTrail(@Nullable ResourceLocation texture) {
-        return texture != null ? QUASAR_TRAIL.apply(texture) : NO_TEXTURE_QUASAR_TRAIL;
+        return quasarTrail(texture, true);
+    }
+
+    public static RenderType quasarTrail(@Nullable ResourceLocation texture, boolean additive) {
+        return texture != null ? QUASAR_TRAIL.apply(texture, additive) : NO_TEXTURE_QUASAR_TRAIL.apply(additive);
     }
 
     public static TransparencyStateShard noTransparencyShard() {