+++ common/src/main/java/foundry/veil/api/quasar/particle/RenderData.java
@@ -81,6 +81,7 @@ public final class RenderData {
         this.prevRotation.set(particle.getRotation());
         this.prevRadius = particle.getRadius();
         this.packedLight = packedLight;
+        this.tickTrails();
     }
 
     @ApiStatus.Internal
@@ -174,8 +175,19 @@ public final class RenderData {
         return this.trails;
     }
 
+    private void tickTrails() {
+        if (this.trails.isEmpty()) {
+            return;
+        }
+
+        for (Trail trail : this.trails) {
+            trail.pushRotatedPoint(new Vec3(this.particle.getPosition().x, this.particle.getPosition().y, this.particle.getPosition().z), new Vec3(this.particle.getRotation()));
+        }
+    }
+
     // TODO move to renderer
-    public void renderTrails(MatrixStack matrixStack, MultiBufferSource bufferSource, Vec3 cameraPos, int packedLight) {
+    @ApiStatus.Internal
+    public void renderTrails(MatrixStack matrixStack, MultiBufferSource bufferSource, Vec3 cameraPos, int packedLight, float partialTicks) {
         if (this.trails.isEmpty()) {
             return;
         }
@@ -183,8 +195,7 @@ public final class RenderData {
         matrixStack.matrixPush();
         matrixStack.translate(-cameraPos.x(), -cameraPos.y(), -cameraPos.z());
         for (Trail trail : this.trails) {
-            trail.pushRotatedPoint(new Vec3(this.prevPosition.x, this.prevPosition.y, this.prevPosition.z), new Vec3(this.prevRotation.x, this.prevRotation.y, this.prevRotation.z));
-            trail.render(matrixStack, bufferSource.getBuffer(VeilRenderType.quasarTrail(trail.getTexture())), packedLight);
+            trail.render(matrixStack, bufferSource.getBuffer(VeilRenderType.quasarTrail(trail.getTexture(), trail.isAdditive())), packedLight, partialTicks, new Vec3(this.particle.getPosition().x, this.particle.getPosition().y, this.particle.getPosition().z), new Vec3(this.particle.getRotation()));
         }
         matrixStack.matrixPop();
     }