+++ changelog.md
@@ -1,3 +1,3 @@
-- Fix Quasar light module having 0 radius with rgb_points ([#156](https://github.com/FoundryMC/Veil/pull/156))
-- Add celeritas and imguimc incompatibilities
-- Fix shader replacement not working
+- Add mutable getters for light data
+- Implement a particle editor for Quasar ([#163](https://github.com/FoundryMC/Veil/pull/163))
+- Improve trails module ([#164](https://github.com/FoundryMC/Veil/pull/164))